Timestamps
Ability to show timestamps and iteration count.

+++ src/output.h
@@ -25,7 +25,7 @@
 struct output_options {
     char *format;
     int interval;       // Seconds
-    int iterations;
+    unsigned long iterations;
     
 };
